move addUser inside user folder
