+++ server/database/query/user/index.js
@@ -2,10 +2,12 @@ const getUserEventById = require('./getUserEventbyId');
 const getAllUserEvent = require('./getAllUserEvent');
 const getUserById = require('./getUserById');
 const deleteUserEventbyId = require('./deleteUserEventbyId');
+const addUser = require('./addUser');
 
 module.exports = {
   getUserById,
   getAllUserEvent,
   getUserEventById,
   deleteUserEventbyId,
+  addUser,
 };
